Update project.md
When working within a sublime-text project, the in-file declarations of the root document do not work. I added a note to this effect.

If this isn't intended behaviour, I'll submit a bug. I'm undecided if it's better as is or to fall back on in-file declarations if the project file doesn't contain the setting.